+++ app/src/main/java/com/bucharest/qurio/data/local/AchievementManager.kt
@@ -4,7 +4,7 @@ import com.bucharest.qurio.data.local.dao.AchievementDao
 import com.bucharest.qurio.data.local.dao.CharacterDao
 import com.bucharest.qurio.data.local.dao.GameSessionDao
 import com.bucharest.qurio.data.local.dao.UserDao
-import com.bucharest.qurio.data.mapper.toEntity
+import com.bucharest.qurio.data.local.mapper.toEntity
 import com.bucharest.qurio.domain.entity.Achievement
 import javax.inject.Inject
 import javax.inject.Singleton
